Some more significant changes to the StorXParser:
Added support for nitrogen concentration data using ISUSFrame.

Removed getSampleDates() and added getFramesMap(), which includes all frames from StorX, CTD, and ISUS frames.

Modified StorXParser.parse() to handle the very last frame in the data file.  Since parse() looks for the 'SAT' pattern to delimit frames, the last frame was getting lost because there was no subsequent 'SAT' after the frame.  parse() now checks to see if the file buffer position is equal to the capacity (it's hit the end) and it creates a frame with the final frameBuffer.

parse() also now handles light and dark binary ISUS frames.

git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@763 9e959af2-1677-4b95-8350-622619c7fe3b