I'm adding the main() function to auth.py.  The function creates a 'location' dictionary with configuration settings for each PacIOOS deployment site for nearshore sensors.  The dictionary includes the following fields per site:
domain - the domain to be used for returning cookies
testURL - the URL used for testing IP connectivity
interval - the interval between tests in seconds
isAuthenticated - defaults to False, set to True to bypass authentication
login - the login service URL and parameters to be sent for authentication
logout - the logout service URL and parameters to be sent to close the session

The main() function creates an instance of the Authentication class, and calls start() to begin the authentication process.  Authentication occurs at the set interval in the configuration dictionary.


git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@774 9e959af2-1677-4b95-8350-622619c7fe3b