I'm adding method stubs for a number of the calibrations defined in Satlantic's instrument File Standard, along with two generic apply() methods with signatures for either Double or String arguments.  these include:
apply() - A method that applies a Satlantic calibration with no coefficients.
apply() - A method that applies a Satlantic calibration to an observed value using the given coefficients.
applyOptic1() - A method that applies a Satlantic OPTIC1 calibration to an observed value using the given coefficients.
applyOptic2() - A method that applies a Satlantic OPTIC2 calibration to an observed value using the given coefficients.
applyOptic3() - A method that applies a Satlantic OPTIC3 calibration to an observed value using the given coefficients.
applyTherm1() - A method that applies a Satlantic THERM1 calibration to an observed value using the given coefficients.
applyPow10() - A method that applies a Satlantic POW10 calibration to an observed value using the given coefficients.
applyPolyU() - A method that applies a Satlantic POLYU calibration to an observed value using the given coefficients.
applyPolyF() - A method that applies a Satlantic POLYF calibration to an observed value using the given coefficients.





git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@814 9e959af2-1677-4b95-8350-622619c7fe3b