Added two new functions to DataProcessor.m:      createPacIOOSFigures()      export2()
The first is the code to create figures for the PacIOOS website, which combine the depth + temperature and salinity + turbidity plots. THe second is a modified version of the export function that will separate the plots into different directories based on date, and can be configured to output in .jpg, .eps, or both.

I also added some code started at 372 to call these functions

Changed processor.m to export all figures through export2().


git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@941 9e959af2-1677-4b95-8350-622619c7fe3b