Iplement the RawToPacIOOS2020Converter methods:
Add a parse() method that parses a series of samples into an internal table structure.
Implement the convert() method that transforms the original samples into a converted table with a UTC datetime column that is sorted chronologically and filters out non-unique rows.
Finish the write() method that writes the converted table to disk.
Add helper methods for getDataPrefix(), setDataPrefix(), getRecordDelimiter(), and setRecordDelimiter().
Change the {get,set}TimeZoneId() methods to use a ZoneId instance rather than a string.

refs https://github.com/csjx/realtime-data/issues/17