Finish out the getArchiver() method:
Handle the dataPrefix character and pass it on to the converter.
Set the channelPath correctly (bug fix).
Set the converter record delimiter to a newline character.
Set the date, time, or datetime formats in the converter based on the sample configuration.
Set the converter column types based on the configuration.

refs https://github.com/csjx/realtime-data/issues/17