Add a TextRebuildApp to rebuild the raw and processed archive:
This application reads in command line parameters of either a directory or file with data samples to rebuild.  It determines the start and end date for the range of samples, and then creates a read task for every file needed to be read within that range.  These are submitted to an executor service, and each task returns a result.  Once all tasks are complete, the data are merged into a single table that is de-duplicated and sorted.  Then given the start and end dates, a set of write tasks are submitted to the executor service, each of which writes 1) a daily processed file and 2) hourly raw files for the given day.  This happens for every day in the range. Once complete, the results are reported.  Original files in the directory are moved to a recovery directory for manual deletion.

refs https://github.com/csjx/realtime-data/issues/20