Upgrade manage-instruments to support rebuilding archives:
Add a rebuilder command that takes a -i instrument argument, and option -f file or -d directory arguments.  This allows rebuilding an archive subset from either a file anywhere on the disk or from a directory that is already part of the archive tree (basically restricted to /data/raw). Rebuilding with recreate both raw and processed archive files.

Use color output for the script feedback.

refs https://github.com/csjx/realtime-data/issues/20