Refactor TextRebuildApp to excise common functionality:
Create a TextRebuilder class that does the heavy lifting of reading, merging, deduplicating, and writing new raw and processed data by refactoring the code out of the TextRebuildApp.  This allows us to use the TextRebuilder in other apps like a TextLoaderApp that will selectively load data into the Data Turbine.

refs https://github.com/csjx/realtime-data/issues/21