Fix the RawToPacIOOS2020Converter timestamp output:
While in the convert() method we called InstantColumn.setPrintFormatter() with a format that excludes milliseconds, that InstantColumn gets overwritten with a new instance from the dateTime column, so the formatter gets lost.

Set the formatter after the InstantColumn is created and populated to ensure milliseconds are removed.  Given the precision of the data, the DateTimeFormatter.ISO_INSTANT pattern works fine.

refs https://github.com/csjx/realtime-data/issues/59